+++ cpp/include/decimal.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <ostream>
+#include <stdint.h>
 #include <string>
 #include <utility>